[github] Use Option<&str> instead of Option<DateTime> for `list_workflow_runs`'s `created` parameter.
The issue is described in more details in https://github.com/github/rest-api-description/issues/2088

For now, this change will treat the created parameter as an optional
str, which it would probably do if the parameter had a custom format
defined.
Alternatively, one oculd try to implement a proper datetime rnage that
follows the [Query for
dates](https://docs.github.com/en/search-github/getting-started-with-searching-on-github/understanding-the-search-syntax#query-for-dates)
format, but this is maybe more work than one would benefit by
implementing the search formating on the user side.