[github] Use specialized error
Most, if not all errors, are generated by, or wrapped by, anyhow!.
This causes all errors to look the same from the caller vantage point, preventing
the caller from possibly handling them. For instance, when the API is
rate limited, returning an error that indicates this is a RateLimiting
error, and providing a simple way to know until when we are being rate
limited is valuable. At the moment, one needs to extract the message from the
anyhow error, do some regex magic to detect if this is being rate
limited or not, and extract until when the rate limitation applies.

This change introduces the use of the crate `thiserror` and ports the
github crate to use it.
All crates will now return a `ClientResult` instead of a
`anyhow::Result`. For crates ported to `thiserror`, Client result is an
alias to `Result<T, ClientError>`, where `ClientError` is the enum
defined using `thiserror` to populate variants.
If a crate is not onboarded in `thiserror`, `ClientResult` is a simple
alias to `anyhow::Result`.