[github] fix potential overflow when computing rate limiting duration
There is a change that `SystemTime::now()` is computed the second after
the "reset" value, or simply clocks could also be slightly off.

When running in "release" mode, this can lead to an overflow, with the
value returned being wrapped to the high u64 values.

https://play.rust-lang.org/?version=stable&mode=release&edition=2021&gist=0c3610ae88babccc0c71ae6e18b07e60

This change ensures that we floor our result to 0 instead of wrapping to
u64::MAX